small bug fix

+++ shoppingcart/Electronics.java
@@ -52,7 +52,7 @@ public class Electronics extends PurchaseItem
 	 * @param isItemFragile Is the item fragile?
 	 * @param itemStateName State the item is being delivered to (for sales tax calculation)
 	 */
-	public Electronics(String itemName, double itemPrice, int itemQuantity, double itemWeight, boolean isItemFragile, String itemStateName) {
+	public Electronics(String itemName, double itemPrice, int itemQuantity, int itemWeight, boolean isItemFragile, String itemStateName) {
 		super(itemName, itemPrice, itemQuantity, itemWeight);
 		isFragile = isItemFragile;
 		stateName = itemStateName;
